organização endpoints
